Fix lint errors and move default options to separate func
